+++ src/apps/vpn/apppermission.cpp
@@ -124,7 +124,7 @@ QVariant AppPermission::data(const QModelIndex& index, int role) const {
 int AppPermission::disabledAppCount() {
   SettingsHolder* settingsHolder = SettingsHolder::instance();
   QStringList applist = settingsHolder->vpnDisabledApps();
-  return applist.count();
+  return static_cast<int>(applist.count());
 }
 
 void AppPermission::flip(const QString& appID) {
