+++ src/apps/vpn/models/servercity.cpp
@@ -113,12 +113,12 @@ const QString ServerCity::localizedName() const {
   return ServerI18N::instance()->translateCityName(m_country, m_name);
 }
 
-unsigned int ServerCity::latency() const {
+qint64 ServerCity::latency() const {
   ServerLatency* serverLatency = MozillaVPN::instance()->serverLatency();
   int numLatencySamples = 0;
-  uint32_t sumLatencyMsec = 0;
+  qint64 sumLatencyMsec = 0;
   for (const QString& pubkey : m_servers) {
-    unsigned int rtt = serverLatency->getLatency(pubkey);
+    qint64 rtt = serverLatency->getLatency(pubkey);
     if (rtt > 0) {
       sumLatencyMsec += rtt;
       numLatencySamples++;
@@ -140,7 +140,7 @@ int ServerCity::connectionScore() const {
   }
 
   // If there is no latency data (zero), then we haven't measured anything yet.
-  unsigned int cityLatencyMsec = latency();
+  qint64 cityLatencyMsec = latency();
   if (cityLatencyMsec == 0) {
     return score;
   }
